Fix memory corruption
The local table implementation assumed that the reader state would
stay alive for at least as long as all the values were used. When
making tests a bit more strict, it surfaced that the blobs were
suddenly corrupted; the backing memory array was deallocated.

We now explicitly create a value-copy of the blob. This does cause a
performance degradation, so perhaps at some point we'll need to
support both use cases.