QDB-10200 - Fix Win64 segmentation fault (#7)
This PR fixes a segfault that occurred on win64, but the bug was present on all platforms. In the JNI native code, there was an implicit assumption that `sizeof(long) == sizeof(jlong) == sizeof(std::uintptr_t)`. On win64, however, this didn't hold.

In various places, rather than casting to `jlong`, we casted to long instead. This caused the first 4 bytes of the point address to be truncated. This didn't surface earlier, because nothing else really overwrote the first 4 bytes. In more recent versions of the JVM, however, this issue did surface on win64.

The PR addresses this by:

* Introducing a "native_ptr" helper module that has functions to (safely) cast from and to JVM;
* Introduce static asserts which verify (at compile time) that the objects we're trying to cast fit in jlong;
* Refactor batch writer and bulk reader functions to make use of these utility functions.

All other functions have been manually audited that they're not doing any invalid casts.

Further work:

* Ensure all functions make use of these helper functions;
* We want to enable -Wall -Werror, as this would have caught these unsafe type casts much earlier.